Split rules.go and regroup tests for maintainability
Extracts Set, Assertion, their construction helpers, validation
pipeline, and MarshalJSON methods into a new set.go. Regroups tests
into field_test.go, set_test.go, and validate_test.go by concern.
