+++ rules/rules_test.go
@@ -1,7 +1,6 @@
 package rules_test
 
 import (
-	"encoding/json"
 	"testing"
 
 	"github.com/invopop/gobl/rules"
@@ -46,13 +45,69 @@ func emailRules() *rules.Set {
 	)
 }
 
+// TestCode is a named string type used to test For with a value type.
+type TestCode string
+
+func testCodeRules() *rules.Set {
+	return rules.For(TestCode(""),
+		rules.Assert("01", "code must not be empty",
+			is.Present,
+		),
+		rules.Assert("02", "code must not exceed 10 characters",
+			is.Expr(`len(this) <= 10`),
+		),
+	)
+}
+
+// Tag is a named string type used as a map key to test map key validation.
+type Tag string
+
+type Tagged struct {
+	Tags map[Tag]Email `json:"tags,omitempty"`
+}
+
+func tagRules() *rules.Set {
+	return rules.For(Tag(""),
+		rules.Assert("01", "tag must not be empty", is.Present),
+	)
+}
+
+// Wrapper and Inner types for Embeddable tests.
+type Inner struct {
+	Name string `json:"name"`
+}
+
+type Wrapper struct {
+	inner *Inner
+}
+
+func (w *Wrapper) Embedded() any {
+	return w.inner
+}
+
+func innerRules() *rules.Set {
+	return rules.For(new(Inner),
+		rules.Field("name",
+			rules.Assert("01", "inner name is required", is.Present),
+		),
+	)
+}
+
+// Container holds a Wrapper in a JSON field to test path prefixing.
+type Container struct {
+	Doc *Wrapper `json:"doc"`
+}
+
 func init() {
 	rules.Register(
 		"test",
 		rules.GOBL.Add("TEST"),
 		emailRules(),
 		personRules(),
+		testCodeRules(),
+		tagRules(),
 	)
+	rules.Register("embeddable-test", rules.GOBL.Add("EMBTEST"), innerRules())
 }
 
 func TestFor(t *testing.T) {
@@ -99,97 +154,6 @@ func TestFor(t *testing.T) {
 	})
 }
 
-func TestValidate(t *testing.T) {
-	t.Run("passes with valid email", func(t *testing.T) {
-		e := &Email{Addr: "test@example.com"}
-		faults := rules.Validate(e)
-		assert.NoError(t, faults)
-	})
-
-	t.Run("fails with blank email address", func(t *testing.T) {
-		e := &Email{Addr: ""}
-		faults := rules.Validate(e)
-		require.NotNil(t, faults)
-		assert.Equal(t, "[GOBL-TEST-EMAIL-01] ($.addr) expected a valid email address", faults.Error())
-		assert.True(t, faults.HasPath("$.addr"))
-		f := faults.First()
-		assert.Equal(t, "expected a valid email address", f.Message())
-		assert.Equal(t, rules.Code("GOBL-TEST-EMAIL-01"), f.Code())
-		assert.Equal(t, []string{"$.addr"}, f.Paths())
-	})
-
-	t.Run("recurses into struct fields", func(t *testing.T) {
-		p := &Person{
-			Name: "Alice",
-			Emails: []Email{
-				{Addr: "ok@example.com"},
-				{Addr: ""},
-			},
-			Address: new(Address),
-		}
-		faults := rules.Validate(p)
-		require.NotNil(t, faults)
-		assert.True(t, faults.HasPath("$.emails[1].addr"), "expected fault at $.emails[1].addr")
-		assert.Equal(t, "[GOBL-TEST-PERSON-01] person address must have a city; [GOBL-TEST-EMAIL-01] ($.emails[1].addr) expected a valid email address", faults.Error())
-	})
-
-	t.Run("recurses into pointer fields", func(t *testing.T) {
-		p := &Person{
-			Name:    "Bob",
-			Address: &Address{Street: "123 Main St", City: "Anytown"},
-		}
-		// Address has no rules, should pass.
-		faults := rules.Validate(p)
-		assert.Nil(t, faults)
-	})
-
-	t.Run("nil pointer field is skipped", func(t *testing.T) {
-		p := &Person{Name: "Carol", Address: &Address{City: "Anytown"}}
-		faults := rules.Validate(p)
-		assert.Nil(t, faults)
-	})
-
-}
-
-// TestCode is a named string type used to test For with a value type.
-type TestCode string
-
-func testCodeRules() *rules.Set {
-	return rules.For(TestCode(""),
-		rules.Assert("01", "code must not be empty",
-			is.Present,
-		),
-		rules.Assert("02", "code must not exceed 10 characters",
-			is.Expr(`len(this) <= 10`),
-		),
-	)
-}
-
-func init() {
-	rules.Register(
-		"test",
-		rules.GOBL.Add("TEST"),
-		testCodeRules(),
-	)
-}
-
-// Tag is a named string type used as a map key to test map key validation.
-type Tag string
-
-type Tagged struct {
-	Tags map[Tag]Email `json:"tags,omitempty"`
-}
-
-func tagRules() *rules.Set {
-	return rules.For(Tag(""),
-		rules.Assert("01", "tag must not be empty", is.Present),
-	)
-}
-
-func init() {
-	rules.Register("test", rules.GOBL.Add("TEST"), tagRules())
-}
-
 func TestForValue(t *testing.T) {
 	t.Run("passes with valid code", func(t *testing.T) {
 		set := testCodeRules()
@@ -233,452 +197,6 @@ func TestForValue(t *testing.T) {
 				),
 			)
 		})
-	})
-
-}
-
-func TestMapValidation(t *testing.T) {
-	t.Run("traverses map values and reports keyed paths", func(t *testing.T) {
-		obj := &Tagged{
-			Tags: map[Tag]Email{
-				"work": {Addr: "work@example.com"},
-				"home": {Addr: ""},
-			},
-		}
-		faults := rules.Validate(obj)
-		require.NotNil(t, faults)
-		assert.True(t, faults.HasPath("$.tags.home.addr"), "expected fault at $.tags.home.addr")
-		assert.False(t, faults.HasPath("$.tags.work.addr"), "expected no fault at $.tags.work.addr")
-	})
-
-	t.Run("validates named map keys", func(t *testing.T) {
-		obj := &Tagged{
-			Tags: map[Tag]Email{
-				"":     {Addr: "ok@example.com"},
-				"work": {Addr: "work@example.com"},
-			},
-		}
-		faults := rules.Validate(obj)
-		require.NotNil(t, faults)
-		assert.True(t, faults.HasPath("$.tags"), "expected fault at $.tags for empty key")
-	})
-
-	t.Run("key order is deterministic across multiple failing entries", func(t *testing.T) {
-		obj := &Tagged{
-			Tags: map[Tag]Email{
-				"beta":  {Addr: ""},
-				"alpha": {Addr: ""},
-			},
-		}
-		faults := rules.Validate(obj)
-		require.NotNil(t, faults)
-		// Faults with same code+message are merged, so both paths appear in one fault
-		assert.True(t, faults.HasPath("$.tags.alpha.addr"))
-		assert.True(t, faults.HasPath("$.tags.beta.addr"))
-	})
-
-	t.Run("nil map is skipped", func(t *testing.T) {
-		obj := &Tagged{Tags: nil}
-		faults := rules.Validate(obj)
-		assert.Nil(t, faults)
-	})
-
-	t.Run("empty map passes", func(t *testing.T) {
-		obj := &Tagged{Tags: map[Tag]Email{}}
-		faults := rules.Validate(obj)
-		assert.Nil(t, faults)
-	})
-}
-
-func TestSetValidate(t *testing.T) {
-	t.Run("skips set when type does not match", func(t *testing.T) {
-		type Other struct{ X string }
-		set := rules.For(new(Email),
-			rules.Assert("01", "always fails",
-				is.Expr(`false`),
-			),
-		)
-		faults := set.Validate(&Other{X: "hello"})
-		assert.Nil(t, faults)
-	})
-
-	t.Run("provides object of same pointer type to tests", func(t *testing.T) {
-		set := rules.For(new(Person),
-			rules.Assert("01", "expected a valid name",
-				is.Func("valid", func(value any) bool {
-					p, ok := value.(*Person)
-					if !ok {
-						return false
-					}
-					return p.Name != ""
-				}),
-			),
-		)
-		p := &Person{Name: "fooo"}
-		faults := set.Validate(p)
-		assert.NoError(t, faults)
-
-		op := Person{Name: "fooo"}
-		faults = set.Validate(op)
-		assert.NoError(t, faults)
-	})
-
-	t.Run("when condition skips set when false", func(t *testing.T) {
-		type Item struct {
-			Active bool   `json:"active"`
-			Name   string `json:"name"`
-		}
-		set := rules.For(new(Item),
-			rules.When(is.Expr(`Active`),
-				rules.Field("name",
-					rules.Assert("01", "name required when active",
-						is.Present,
-					),
-				),
-			),
-		)
-		// Active is false: set should be skipped.
-		faults := set.Validate(&Item{Active: false, Name: ""})
-		assert.NoError(t, faults)
-	})
-
-	t.Run("when condition runs assertions when true", func(t *testing.T) {
-		type Item struct {
-			Active bool   `json:"active"`
-			Name   string `json:"name"`
-		}
-		set := rules.For(new(Item),
-			rules.When(is.Expr(`Active`),
-				rules.Field("name",
-					rules.Assert("01", "name required when active",
-						is.Present,
-					),
-				),
-			),
-		)
-		// Active is true, Name is blank: should fail.
-		faults := set.Validate(&Item{Active: true, Name: ""})
-		require.Error(t, faults)
-	})
-
-	t.Run("when condition runs assertions when true with expr", func(t *testing.T) {
-		type Item struct {
-			Active bool   `json:"active"`
-			Name   string `json:"name"`
-		}
-		proto := new(Item)
-		set := rules.For(proto,
-			rules.When(is.Expr(`Active`),
-				rules.Assert("01", "name required when active",
-					is.Expr(`Name != ""`),
-				),
-			),
-		)
-		// Active is true, Name is blank: should fail.
-		faults := set.Validate(&Item{Active: true, Name: ""})
-		require.Error(t, faults)
-	})
-
-}
-
-func TestFieldEmpty(t *testing.T) {
-	set := rules.For(new(Person),
-		rules.Field("name",
-			rules.Assert("01", "name must not be set", is.Empty),
-		),
-	)
-
-	t.Run("passes when field is empty string", func(t *testing.T) {
-		faults := set.Validate(&Person{Name: ""})
-		assert.NoError(t, faults)
-	})
-
-	t.Run("fails when field has a value", func(t *testing.T) {
-		faults := set.Validate(&Person{Name: "Alice"})
-		require.Error(t, faults)
-		assert.True(t, faults.HasPath("$.name"))
-	})
-}
-
-func TestFieldNil(t *testing.T) {
-	set := rules.For(new(Person),
-		rules.Field("second_address",
-			rules.Assert("01", "second address must not be set", is.Nil),
-		),
-	)
-
-	t.Run("passes when pointer field is nil", func(t *testing.T) {
-		faults := set.Validate(&Person{SecondAddress: nil})
-		assert.NoError(t, faults)
-	})
-
-	t.Run("fails when pointer field is non-nil", func(t *testing.T) {
-		faults := set.Validate(&Person{SecondAddress: &Address{City: "London"}})
-		require.Error(t, faults)
-		assert.True(t, faults.HasPath("$.second_address"))
-	})
-
-	t.Run("fails when pointer field is non-nil but points to empty value", func(t *testing.T) {
-		faults := set.Validate(&Person{SecondAddress: new(Address)})
-		require.Error(t, faults)
-		assert.True(t, faults.HasPath("$.second_address"))
-	})
-}
-
-func TestFieldRulesDoNotBleedToSameType(t *testing.T) {
-	// Verifies that rules defined inside a When+Field block for a specific field
-	// (e.g. "address") are NOT applied to other fields of the same type
-	// (e.g. "second_address"). This guards against the namespace traversal
-	// incorrectly matching field subsets by type rather than by field name.
-	set := rules.For(new(Person),
-		rules.When(
-			is.Expr(`Name != ""`),
-			rules.Field("address",
-				rules.Assert("01", "city is required",
-					is.Func("city required", func(val any) bool {
-						a, ok := val.(*Address)
-						return !ok || a == nil || a.City != ""
-					}),
-				),
-			),
-		),
-	)
-
-	t.Run("rule is not applied to second_address field of same type", func(t *testing.T) {
-		p := &Person{
-			Name:          "Alice",
-			Address:       &Address{City: "London"}, // valid
-			SecondAddress: &Address{City: ""},       // would fail city check if rule bled
-		}
-		faults := set.Validate(p)
-		assert.NoError(t, faults)
-	})
-
-	t.Run("rule still fails when the scoped address field fails", func(t *testing.T) {
-		p := &Person{
-			Name:          "Alice",
-			Address:       &Address{City: ""},       // should fail
-			SecondAddress: &Address{City: "London"}, // should not be checked
-		}
-		faults := set.Validate(p)
-		require.Error(t, faults)
-		assert.True(t, faults.HasPath("$.address"))
-		assert.False(t, faults.HasPath("$.second_address"))
-	})
-}
-
-func TestFieldNested(t *testing.T) {
-	set := rules.For(new(Person),
-		rules.Field("address",
-			rules.Field("city",
-				rules.Assert("01", "city is required", is.Present),
-			),
-			rules.Field("street",
-				rules.Assert("02", "street is required", is.Present),
-			),
-		),
-	)
-
-	t.Run("reports fault at nested field path", func(t *testing.T) {
-		p := &Person{Address: &Address{Street: "1 Main", City: ""}}
-		faults := set.Validate(p)
-		require.Error(t, faults)
-		assert.True(t, faults.HasPath("$.address.city"))
-	})
-
-	t.Run("passes when nested fields are set", func(t *testing.T) {
-		p := &Person{Address: &Address{Street: "1 Main", City: "London"}}
-		assert.NoError(t, set.Validate(p))
-	})
-
-	t.Run("skips nested field subset when parent pointer is nil", func(t *testing.T) {
-		p := &Person{Address: nil}
-		assert.NoError(t, set.Validate(p))
-	})
-
-	t.Run("sibling nested fields are independent", func(t *testing.T) {
-		p := &Person{Address: &Address{Street: "1 Main", City: ""}}
-		faults := set.Validate(p)
-		require.Error(t, faults)
-		assert.True(t, faults.HasPath("$.address.city"))
-		assert.False(t, faults.HasPath("$.address.street"))
-	})
-}
-
-func TestEmbeddedStruct(t *testing.T) {
-	type Individual struct {
-		Person
-		Testing bool `json:"testing"`
-	}
-	set := rules.For(new(Individual),
-		rules.Field("address",
-			rules.Field("city",
-				rules.Assert("01", "city is required", is.Present),
-			),
-		),
-		rules.Field("testing",
-			rules.Assert("02", "testing must be true", is.In(true)),
-		),
-	)
-	t.Run("embedded struct fields are validated", func(t *testing.T) {
-		i := new(Individual)
-		i.Address = &Address{Street: "1 Main", City: "Foo"}
-		i.Testing = true
-		assert.NoError(t, set.Validate(i))
-	})
-}
-
-func TestEach(t *testing.T) {
-	t.Run("validates each element and reports indexed paths", func(t *testing.T) {
-		set := rules.For(new(Person),
-			rules.Field("emails",
-				rules.Each(
-					rules.Field("addr",
-						rules.Assert("01", "email address is required", is.Present),
-					),
-				),
-			),
-		)
-		p := &Person{
-			Name: "Alice",
-			Emails: []Email{
-				{Addr: "a@example.com"},
-				{Addr: ""},
-				{Addr: "b@example.com"},
-				{Addr: ""},
-			},
-		}
-		faults := set.Validate(p)
-		require.Error(t, faults)
-		assert.True(t, faults.HasPath("$.emails[1].addr"))
-		assert.True(t, faults.HasPath("$.emails[3].addr"))
-		assert.False(t, faults.HasPath("$.emails[0].addr"))
-		assert.False(t, faults.HasPath("$.emails[2].addr"))
-	})
-
-	t.Run("passes when all elements are valid", func(t *testing.T) {
-		set := rules.For(new(Person),
-			rules.Field("emails",
-				rules.Each(
-					rules.Field("addr",
-						rules.Assert("01", "email address is required", is.Present),
-					),
-				),
-			),
-		)
-		p := &Person{
-			Emails: []Email{{Addr: "a@example.com"}, {Addr: "b@example.com"}},
-		}
-		assert.NoError(t, set.Validate(p))
-	})
-
-	t.Run("passes with empty slice", func(t *testing.T) {
-		set := rules.For(new(Person),
-			rules.Field("emails",
-				rules.Each(
-					rules.Assert("01", "required", is.Present),
-				),
-			),
-		)
-		assert.NoError(t, set.Validate(&Person{}))
-	})
-
-	t.Run("whole-slice and per-element assertions coexist on same field", func(t *testing.T) {
-		set := rules.For(new(Person),
-			rules.Field("emails",
-				rules.Assert("01", "no more than two emails",
-					is.Func("max two", func(val any) bool {
-						emails, ok := val.([]Email)
-						return !ok || len(emails) <= 2
-					}),
-				),
-				rules.Each(
-					rules.Field("addr",
-						rules.Assert("02", "email address is required", is.Present),
-					),
-				),
-			),
-		)
-		// Whole-slice violation: three emails.
-		p := &Person{Emails: []Email{{Addr: "a@b.com"}, {Addr: "c@d.com"}, {Addr: "e@f.com"}}}
-		faults := set.Validate(p)
-		require.Error(t, faults)
-		assert.True(t, faults.HasPath("$.emails"))
-
-		// Per-element violation: empty addr.
-		p2 := &Person{Emails: []Email{{Addr: "a@b.com"}, {Addr: ""}}}
-		faults2 := set.Validate(p2)
-		require.Error(t, faults2)
-		assert.True(t, faults2.HasPath("$.emails[1].addr"))
-	})
-
-	t.Run("panics when used outside a slice field", func(t *testing.T) {
-		assert.Panics(t, func() {
-			rules.For(new(Email),
-				rules.Each(
-					rules.Assert("01", "required", is.Present),
-				),
-			)
-		})
-	})
-}
-
-// Wrapper and Inner types for Embeddable tests.
-type Inner struct {
-	Name string `json:"name"`
-}
-
-type Wrapper struct {
-	inner *Inner
-}
-
-func (w *Wrapper) Embedded() any {
-	return w.inner
-}
-
-func innerRules() *rules.Set {
-	return rules.For(new(Inner),
-		rules.Field("name",
-			rules.Assert("01", "inner name is required", is.Present),
-		),
-	)
-}
-
-// Container holds a Wrapper in a JSON field to test path prefixing.
-type Container struct {
-	Doc *Wrapper `json:"doc"`
-}
-
-func init() {
-	rules.Register("embeddable-test", rules.GOBL.Add("EMBTEST"), innerRules())
-}
-
-func TestEmbeddable(t *testing.T) {
-	t.Run("faults from embedded payload carry no extra prefix", func(t *testing.T) {
-		w := &Wrapper{inner: &Inner{Name: ""}}
-		set := innerRules()
-		faults := set.Validate(w)
-		// The set is scoped to Inner; Wrapper itself has no matching rules.
-		assert.Nil(t, faults)
-	})
-
-	t.Run("global Validate traverses Embeddable and prefixes path correctly", func(t *testing.T) {
-		c := &Container{Doc: &Wrapper{inner: &Inner{Name: ""}}}
-		faults := rules.Validate(c)
-		require.Error(t, faults)
-		assert.True(t, faults.HasPath("$.doc.name"), "expected fault at $.doc.name, got: %v", faults)
-	})
-
-	t.Run("global Validate passes when embedded payload is valid", func(t *testing.T) {
-		c := &Container{Doc: &Wrapper{inner: &Inner{Name: "Alice"}}}
-		faults := rules.Validate(c)
-		assert.Nil(t, faults)
-	})
-
-	t.Run("nil embedded payload is skipped", func(t *testing.T) {
-		c := &Container{Doc: &Wrapper{inner: nil}}
-		faults := rules.Validate(c)
-		assert.Nil(t, faults)
 	})
 }
 
@@ -759,337 +277,4 @@ func TestAllSets(t *testing.T) {
 	for _, s := range reg {
 		assert.Contains(t, sets, s)
 	}
-}
-
-func TestSetMarshalJSON(t *testing.T) {
-	set := rules.For(new(Person),
-		rules.When(is.Expr(`Age > 0`),
-			rules.Assert("01", "name required",
-				is.Expr(`Name != ""`),
-			),
-		),
-		rules.Field("name",
-			rules.Assert("02", "name not empty", is.Present),
-		),
-	)
-	data, err := json.Marshal(set)
-	require.NoError(t, err)
-
-	var m map[string]any
-	require.NoError(t, json.Unmarshal(data, &m))
-	assert.Contains(t, m, "id")
-	assert.Contains(t, m, "subsets")
-}
-
-func TestAssertionMarshalJSON(t *testing.T) {
-	set := rules.For(new(Person),
-		rules.Assert("01", "name and age valid",
-			is.Expr(`Name != ""`),
-			is.Expr(`Age > 0`),
-		),
-	)
-	data, err := json.Marshal(set)
-	require.NoError(t, err)
-
-	var m map[string]any
-	require.NoError(t, json.Unmarshal(data, &m))
-	asserts, ok := m["assert"].([]any)
-	require.True(t, ok)
-	require.Len(t, asserts, 1)
-	a := asserts[0].(map[string]any)
-	// Tests should be comma-joined string
-	tests, ok := a["tests"].(string)
-	require.True(t, ok)
-	assert.Contains(t, tests, ", ")
-}
-
-func TestValidateNil(t *testing.T) {
-	faults := rules.Validate(nil)
-	assert.Nil(t, faults)
-}
-
-// testRegime implements ContextAdder for guard testing.
-type testRegime struct {
-	Code string
-}
-
-func (r testRegime) RulesContext() rules.WithContext {
-	return func(rc *rules.Context) {
-		rc.Set("regime", r.Code)
-	}
-}
-
-type docWithRegime struct {
-	Regime testRegime `json:"regime"`
-	Name   string     `json:"name"`
-}
-
-func TestRegisterWithGuard(t *testing.T) {
-	guardTest := is.In("ES", "PT")
-	nameSet := rules.For(new(docWithRegime),
-		rules.Field("name",
-			rules.Assert("01", "name required", is.Present),
-		),
-	)
-
-	rules.RegisterWithGuard("guard-test", rules.GOBL.Add("GUARDTEST"),
-		is.InContext(guardTest), nameSet)
-
-	t.Run("guard passes - set applied", func(t *testing.T) {
-		doc := &docWithRegime{
-			Regime: testRegime{Code: "ES"},
-			Name:   "",
-		}
-		faults := rules.Validate(doc)
-		require.Error(t, faults)
-		assert.True(t, faults.HasCode("GOBL-GUARDTEST-DOCWITHREGIME-01"))
-	})
-
-	t.Run("guard fails - set skipped", func(t *testing.T) {
-		doc := &docWithRegime{
-			Regime: testRegime{Code: "FR"},
-			Name:   "",
-		}
-		faults := rules.Validate(doc)
-		// Should not have the guard-test fault (guard doesn't match FR)
-		if faults != nil {
-			assert.False(t, faults.HasCode("GOBL-GUARDTEST-DOCWITHREGIME-01"))
-		}
-	})
-}
-
-func TestAssertIfPresentMarshalJSON(t *testing.T) {
-	// This triggers presentGuard.String() via Set.MarshalJSON.
-	set := rules.For(new(Person),
-		rules.Field("name",
-			rules.AssertIfPresent("01", "name too short",
-				is.Func("min 3", func(val any) bool {
-					s, ok := val.(string)
-					return ok && len(s) >= 3
-				}),
-			),
-		),
-	)
-	data, err := json.Marshal(set)
-	require.NoError(t, err)
-	assert.Contains(t, string(data), "present")
-}
-
-func TestValidateWithContext(t *testing.T) {
-	// Tests the WithContext option path in Validate.
-	guardTest := is.In("DE")
-	nameSet := rules.For(new(Email),
-		rules.Field("addr",
-			rules.Assert("01", "email required", is.Present),
-		),
-	)
-	rules.RegisterWithGuard("ctx-opt-test", rules.GOBL.Add("CTXOPT"),
-		is.InContext(guardTest), nameSet)
-
-	t.Run("WithContext option injects context", func(t *testing.T) {
-		e := &Email{Addr: ""}
-		faults := rules.Validate(e, func(rc *rules.Context) {
-			rc.Set("country", "DE")
-		})
-		require.Error(t, faults)
-		assert.True(t, faults.HasCode("GOBL-CTXOPT-EMAIL-01"))
-	})
-}
-
-func TestEachWithPointerElements(t *testing.T) {
-	type Item struct {
-		Value string `json:"value"`
-	}
-	type Container struct {
-		Items []*Item `json:"items"`
-	}
-
-	set := rules.For(new(Container),
-		rules.Field("items",
-			rules.Each(
-				rules.Field("value",
-					rules.Assert("01", "value required", is.Present),
-				),
-			),
-		),
-	)
-
-	t.Run("pointer elements validated", func(t *testing.T) {
-		c := &Container{Items: []*Item{{Value: "ok"}, {Value: ""}}}
-		faults := set.Validate(c)
-		require.Error(t, faults)
-		assert.True(t, faults.HasPath("$.items[1].value"))
-	})
-
-	t.Run("all valid", func(t *testing.T) {
-		c := &Container{Items: []*Item{{Value: "a"}, {Value: "b"}}}
-		faults := set.Validate(c)
-		assert.NoError(t, faults)
-	})
-
-	t.Run("nil slice field", func(t *testing.T) {
-		c := &Container{Items: nil}
-		faults := set.Validate(c)
-		assert.NoError(t, faults)
-	})
-}
-
-func TestEachTopLevelSlice(t *testing.T) {
-	// Tests validateEachValue with pointer-to-slice and Each at top-level of When.
-	type Row struct {
-		Name string `json:"name"`
-	}
-	type Table struct {
-		Rows []Row `json:"rows"`
-	}
-	set := rules.For(new(Table),
-		rules.Field("rows",
-			rules.Each(
-				rules.Assert("01", "name required",
-					is.Func("has name", func(val any) bool {
-						r, ok := val.(*Row)
-						return ok && r.Name != ""
-					}),
-				),
-			),
-		),
-	)
-	t.Run("each validates all rows", func(t *testing.T) {
-		tb := &Table{Rows: []Row{{Name: "a"}, {Name: ""}}}
-		faults := set.Validate(tb)
-		require.Error(t, faults)
-		assert.True(t, faults.HasPath("$.rows[1]"))
-	})
-}
-
-func TestValidateWithNilPointerField(_ *testing.T) {
-	// Nil pointer field should be skipped without panic.
-	p := &Person{Name: "Alice", Address: nil}
-	faults := rules.Validate(p)
-	// No rules fire since address is nil and name is set.
-	_ = faults
-}
-
-type noJSONTag struct {
-	Bare string
-}
-
-func TestFieldWithNoJSONTag(t *testing.T) {
-	// Tests the jsonFieldName fallback to field name when no json tag.
-	set := rules.For(new(noJSONTag),
-		rules.Field("Bare",
-			rules.Assert("01", "bare required", is.Present),
-		),
-	)
-	faults := set.Validate(&noJSONTag{Bare: ""})
-	require.Error(t, faults)
-	assert.True(t, faults.HasPath("$.Bare"))
-}
-
-type dashTag struct {
-	Hidden string `json:"-"`
-	Shown  string `json:"shown"`
-}
-
-func TestFieldWithDashJSONTag(t *testing.T) {
-	// Exercises the json:"-" path in jsonFieldName.
-	set := rules.For(new(dashTag),
-		rules.Field("shown",
-			rules.Assert("01", "shown required", is.Present),
-		),
-	)
-	faults := set.Validate(&dashTag{Shown: ""})
-	require.Error(t, faults)
-}
-
-func TestNewSet(t *testing.T) {
-	t.Run("codes are namespaced", func(t *testing.T) {
-		// Use top-level helper so runtime.Caller(1) in For detects the
-		// correct package and omits the package prefix from the set ID.
-		set := rules.NewSet("MYAPP", emailRules())
-		faults := set.Validate(&Email{Addr: ""})
-		require.Error(t, faults)
-		assert.Equal(t, rules.Code("MYAPP-EMAIL-01"), faults.First().Code())
-		assert.True(t, faults.HasPath("$.addr"))
-	})
-
-	t.Run("multi-type validation", func(t *testing.T) {
-		set := rules.NewSet("NS", personRules(), emailRules())
-
-		// Validates Person with nested Email via type index traversal.
-		p := &Person{
-			Name:    "Alice",
-			Address: new(Address),
-			Emails: []Email{
-				{Addr: "ok@example.com"},
-				{Addr: ""},
-			},
-		}
-		faults := set.Validate(p)
-		require.Error(t, faults)
-		assert.True(t, faults.HasCode("NS-PERSON-01"))
-		assert.True(t, faults.HasCode("NS-EMAIL-01"))
-		assert.True(t, faults.HasPath("$.emails[1].addr"))
-	})
-
-	t.Run("does not pollute global registry", func(t *testing.T) {
-		before := len(rules.AllSets())
-		rules.NewSet("STANDALONE", emailRules())
-		after := len(rules.AllSets())
-		assert.Equal(t, before, after)
-	})
-
-	t.Run("passes when valid", func(t *testing.T) {
-		set := rules.NewSet("OK", emailRules())
-		faults := set.Validate(&Email{Addr: "test@example.com"})
-		assert.NoError(t, faults)
-	})
-
-	t.Run("same For output reused across multiple NewSet calls", func(t *testing.T) {
-		shared := emailRules()
-		a := rules.NewSet("AAA", shared)
-		b := rules.NewSet("BBB", shared)
-
-		fa := a.Validate(&Email{Addr: ""})
-		fb := b.Validate(&Email{Addr: ""})
-		require.Error(t, fa)
-		require.Error(t, fb)
-		assert.Equal(t, rules.Code("AAA-EMAIL-01"), fa.First().Code())
-		assert.Equal(t, rules.Code("BBB-EMAIL-01"), fb.First().Code())
-		// Original set is untouched.
-		assert.Equal(t, rules.Code("EMAIL"), shared.ID)
-	})
-}
-
-func docWithRegimeRules() *rules.Set {
-	return rules.For(new(docWithRegime),
-		rules.When(is.InContext(is.In("ES")),
-			rules.Field("name",
-				rules.Assert("01", "name required", is.Present),
-			),
-		),
-	)
-}
-
-func TestNewSetValidateWithContext(t *testing.T) {
-	set := rules.NewSet("CTX", docWithRegimeRules())
-
-	t.Run("context-aware guard fires with matching context", func(t *testing.T) {
-		doc := &docWithRegime{
-			Regime: testRegime{Code: "ES"},
-			Name:   "",
-		}
-		faults := set.Validate(doc)
-		require.Error(t, faults)
-		assert.True(t, faults.HasCode("CTX-DOCWITHREGIME-01"))
-	})
-
-	t.Run("context-aware guard skips with non-matching context", func(t *testing.T) {
-		doc := &docWithRegime{
-			Regime: testRegime{Code: "FR"},
-			Name:   "",
-		}
-		faults := set.Validate(doc)
-		assert.NoError(t, faults)
-	})
 }
